Refactor to not use file extension
